Changed hoc to custom hook for change alert modal
