+++ src/ChangeAlert/index.js
@@ -1,7 +1,9 @@
 import React from 'react'
-import { withStorageListener } from './withStorageListener'
+import { useStorageListener } from './useStorageListener'
 import './ChangeAlert.css'
-function ChangeAlert({ show, toggleShow }) {
+function ChangeAlert({ sync }) {
+	console.log(sync)
+	const { show, toggleShow } = useStorageListener({ sync })
 	if (show) {
 		return (
 			<div className="ChangeAlert-bg">
@@ -22,6 +24,4 @@ function ChangeAlert({ show, toggleShow }) {
 	}
 }
 
-const ChangeAlertWithStorageListener = withStorageListener(ChangeAlert)
-
-export { ChangeAlertWithStorageListener }
+export { ChangeAlert }
