--- src/ChangeAlert/withStorageListener.js
@@ -1,26 +1,0 @@
-import React, { useState } from 'react'
-
-function withStorageListener(WrappedComponent) {
-	return function WrappedComponentWithStorageListener(props) {
-		const [storageChange, setStorageChange] = useState(false)
-		React.useEffect(() => {
-			const onChange = (change) => {
-				if (change.key === 'TODOS_V1') {
-					console.log('Hubo cambios en TODOS_V1')
-					setStorageChange(true)
-				}
-			}
-			window.addEventListener('storage', onChange)
-			return () => {
-				window.removeEventListener('storage', onChange)
-			}
-		}, [])
-		const toggleShow = () => {
-			props.sync()
-			setStorageChange(false)
-		}
-		return <WrappedComponent show={storageChange} toggleShow={toggleShow} />
-	}
-}
-
-export { withStorageListener }
